refactor(components): group animation components
