+++ src/common/components/animation/ClipWord/index.tsx
