+++ src/common/components/animation/ClipWord/styles.ts
