+++ src/common/components/animation/FadeInBlock/styles.ts
