+++ src/common/components/animation/FadeInBlock/types.ts
