+++ src/common/components/animation/RevealedText/index.tsx
